initial refactor (wip)
