+++ docs/content/3.api/4.advanced/2.kit.md
@@ -74,7 +74,6 @@ description: Nuxt Kit provides composable utilities to help interacting with Nux
 
 [source code](https://github.com/nuxt/framework/blob/main/packages/kit/src/nitro.ts)
 
-- ~~`addServerMiddleware(serverMiddleware)`~~
 - `addServerHandler (handler)`
 - `addDevServerHandler (handler)`
 - `useNitro()` (only usable after `ready` hook)
