+++ packages/schema/src/config/typescript.ts
@@ -4,8 +4,6 @@ export default defineUntypedSchema({
   /**
    * Configuration for Nuxt's TypeScript integration.
    *
-   * @version 2
-   * @version 3
    */
   typescript: {
     /**
@@ -38,7 +36,8 @@ export default defineUntypedSchema({
     /**
      * Generate a `*.vue` shim.
      *
-     * We recommend instead either enabling [**Take Over Mode**](https://vuejs.org/guide/typescript/overview.html#volar-takeover-mode) or adding **TypeScript Vue Plugin (Volar)** 👉 [[Download](https://marketplace.visualstudio.com/items?itemName=Vue.vscode-typescript-vue-plugin)].
+     * We recommend instead either enabling [**Take Over Mode**](https://vuejs.org/guide/typescript/overview.html#volar-takeover-mode) or adding
+     * TypeScript Vue Plugin (Volar)** 👉 [[Download](https://marketplace.visualstudio.com/items?itemName=Vue.vscode-typescript-vue-plugin)].
      */
     shim: true
   }
