+++ packages/schema/src/nuxt2/config/app.ts
@@ -1,22 +1,15 @@
+import { defineUntypedSchema } from 'untyped'
 import { resolve, join } from 'pathe'
 import { existsSync, readdirSync } from 'node:fs'
 import defu from 'defu'
-import { defineUntypedSchema } from 'untyped'
-import type { AppHeadMetaObject } from '../types/meta'
 
 export default defineUntypedSchema({
-  /**
-   * Vue.js config
-   * @version 2
-   * @version 3
-   */
   vue: {
     /**
      * Properties that will be set directly on `Vue.config` for vue@2.
      *
      * @see [vue@2 Documentation](https://v2.vuejs.org/v2/api/#Global-Config)
      * @type {typeof import('vue/types/vue').VueConfiguration}
-     * @version 2
      */
     config: {
       silent: {
@@ -26,151 +19,18 @@ export default defineUntypedSchema({
         $resolve: async (val, get) => val ?? await get('dev')
       },
     },
-    /**
-     * Options for the Vue compiler that will be passed at build time.
-     * @see [documentation](https://vuejs.org/api/application.html#app-config-compileroptions)
-     * @type {typeof import('@vue/compiler-core').CompilerOptions}
-     * @version 3
-     */
-    compilerOptions: {}
   },
 
-  /**
-   * Nuxt App configuration.
-   * @version 2
-   * @version 3
-   */
   app: {
-    /**
-     * The base path of your Nuxt application.
-     *
-     * This can be set at runtime by setting the NUXT_APP_BASE_URL environment variable.
-     * @example
-     * ```bash
-     * NUXT_APP_BASE_URL=/prefix/ node .output/server/index.mjs
-     * ```
-     */
-    baseURL: {
-      $resolve: async (val) => val || process.env.NUXT_APP_BASE_URL || '/',
-    },
-
-    /** The folder name for the built site assets, relative to `baseURL` (or `cdnURL` if set). This is set at build time and should not be customized at runtime. */
-    buildAssetsDir: {
-      $resolve: async (val) => val || process.env.NUXT_APP_BUILD_ASSETS_DIR || '/_nuxt/',
-    },
-
     /**
      * The folder name for the built site assets, relative to `baseURL` (or `cdnURL` if set).
      * @deprecated - use `buildAssetsDir` instead
-     * @version 2
-     */
+    */
     assetsPath: {
       $resolve: async (val, get) => val ?? (await get('buildAssetsDir'))
     },
-    /**
-     * An absolute URL to serve the public folder from (production-only).
-     *
-     * This can be set to a different value at runtime by setting the `NUXT_APP_CDN_URL` environment variable.
-     * @example
-     * ```bash
-     * NUXT_APP_CDN_URL=https://mycdn.org/ node .output/server/index.mjs
-     * ```
-     */
-    cdnURL: {
-      $resolve: async (val, get) => (await get('dev')) ? '' : (process.env.NUXT_APP_CDN_URL ?? val) || ''
-    },
-    /**
-     * Set default configuration for `<head>` on every page.
-     *
-     * @example
-     * ```js
-     * app: {
-     *   head: {
-     *     meta: [
-     *       // <meta name="viewport" content="width=device-width, initial-scale=1">
-     *       { name: 'viewport', content: 'width=device-width, initial-scale=1' }
-     *     ],
-     *     script: [
-     *       // <script src="https://myawesome-lib.js"></script>
-     *       { src: 'https://awesome-lib.js' }
-     *     ],
-     *     link: [
-     *       // <link rel="stylesheet" href="https://myawesome-lib.css">
-     *       { rel: 'stylesheet', href: 'https://awesome-lib.css' }
-     *     ],
-     *     // please note that this is an area that is likely to change
-     *     style: [
-     *       // <style type="text/css">:root { color: red }</style>
-     *       { children: ':root { color: red }', type: 'text/css' }
-     *     ],
-     *     noscript: [
-     *       // <noscript>Javascript is required</noscript>
-     *       { children: 'Javascript is required' }
-     *     ]
-     *   }
-     * }
-     * ```
-     * @type {typeof import('../src/types/config').NuxtAppConfig['head']}
-     * @version 3
-     */
-    head: {
-      $resolve: async (val, get) => {
-        const resolved: Required<AppHeadMetaObject> = defu(val, await get('meta'), {
-          meta: [],
-          link: [],
-          style: [],
-          script: [],
-          noscript: []
-        })
-
-        // provides default charset and viewport if not set
-        if (!resolved.meta.find(m => m.charset)?.charset) {
-          resolved.meta.unshift({ charset: resolved.charset || 'utf-8' })
-        }
-        if (!resolved.meta.find(m => m.name === 'viewport')?.content) {
-          resolved.meta.unshift({ name: 'viewport', content: resolved.viewport || 'width=device-width, initial-scale=1' })
-        }
-
-        resolved.meta = resolved.meta.filter(Boolean)
-        resolved.link = resolved.link.filter(Boolean)
-        resolved.style = resolved.style.filter(Boolean)
-        resolved.script = resolved.script.filter(Boolean)
-        resolved.noscript = resolved.noscript.filter(Boolean)
-
-        return resolved
-      }
-    },
-    /**
-     * Default values for layout transitions.
-     *
-     * This can be overridden with `definePageMeta` on an individual page.
-     * Only JSON-serializable values are allowed.
-     *
-     * @see https://vuejs.org/api/built-in-components.html#transition
-     * @type {typeof import('../src/types/config').NuxtAppConfig['layoutTransition']}
-     */
-    layoutTransition: false,
-    /**
-     * Default values for page transitions.
-     *
-     * This can be overridden with `definePageMeta` on an individual page.
-     * Only JSON-serializable values are allowed.
-     *
-     * @see https://vuejs.org/api/built-in-components.html#transition
-     * @type {typeof import('../src/types/config').NuxtAppConfig['pageTransition']}
-     */
-    pageTransition: false,
-    /**
-     * Default values for KeepAlive configuration between pages.
-     *
-     * This can be overridden with `definePageMeta` on an individual page.
-     * Only JSON-serializable values are allowed.
-     *
-     * @see https://vuejs.org/api/built-in-components.html#keepalive
-     * @type {typeof import('../src/types/config').NuxtAppConfig['keepalive']}
-     */
-    keepalive: false,
   },
+
   /**
    * The path to an HTML template file for rendering Nuxt responses.
    * Uses `<srcDir>/app.html` if it exists, or the Nuxt's default template if not.
@@ -187,7 +47,6 @@ export default defineUntypedSchema({
    *   </body>
    * </html>
    * ```
-   * @version 2
    */
   appTemplatePath: {
     $resolve: async (val, get) => {
@@ -205,7 +64,6 @@ export default defineUntypedSchema({
    * Enable or disable Vuex store.
    *
    * By default, it is enabled if there is a `store/` directory.
-   * @version 2
    */
   store: {
     $resolve: async (val, get) => val !== false &&
@@ -219,7 +77,6 @@ export default defineUntypedSchema({
    *
    * @see [documentation](https://vue-meta.nuxtjs.org/api/#plugin-options).
    * @type {typeof import('vue-meta').VueMetaOptions}
-   * @version 2
    */
   vueMeta: null,
 
@@ -228,7 +85,6 @@ export default defineUntypedSchema({
    *
    * @see [documentation](https://vue-meta.nuxtjs.org/api/#metainfo-properties) for specifics.
    * @type {typeof import('vue-meta').MetaInfo}
-   * @version 2
    */
   head: {
     /** Each item in the array maps to a newly-created `<meta>` element, where object properties map to attributes. */
@@ -243,7 +99,6 @@ export default defineUntypedSchema({
 
   /**
    * @type {typeof import('../src/types/meta').AppHeadMetaObject}
-   * @version 3
    * @deprecated - use `head` instead
    */
   meta: {
@@ -255,7 +110,6 @@ export default defineUntypedSchema({
 
   /**
    * Configuration for the Nuxt `fetch()` hook.
-   * @version 2
    */
   fetch: {
     /** Whether to enable `fetch()` on the server. */
@@ -265,72 +119,18 @@ export default defineUntypedSchema({
   },
 
   /**
-   * An array of nuxt app plugins.
-   *
-   * Each plugin can be a string (which can be an absolute or relative path to a file).
-   * If it ends with `.client` or `.server` then it will be automatically loaded only
-   * in the appropriate context.
-   *
-   * It can also be an object with `src` and `mode` keys.
-   *
-   * @example
-   * ```js
-   * plugins: [
-   *   '~/plugins/foo.client.js', // only in client side
-   *   '~/plugins/bar.server.js', // only in server side
-   *   '~/plugins/baz.js', // both client & server
-   *   { src: '~/plugins/both-sides.js' },
-   *   { src: '~/plugins/client-only.js', mode: 'client' }, // only on client side
-   *   { src: '~/plugins/server-only.js', mode: 'server' } // only on server side
-   * ]
-   * ```
-   * @type {(typeof import('../src/types/nuxt').NuxtPlugin | string)[]}
-   * @version 2
-   */
-  plugins: [],
-
-  /**
    * You may want to extend plugins or change their order. For this, you can pass
    * a function using `extendPlugins`. It accepts an array of plugin objects and
    * should return an array of plugin objects.
    * @type {(plugins: Array<{ src: string, mode?: 'client' | 'server' }>) => Array<{ src: string, mode?: 'client' | 'server' }>}
-   * @version 2
    */
   extendPlugins: null,
-
-  /**
-   * You can define the CSS files/modules/libraries you want to set globally
-   * (included in every page).
-   *
-   * Nuxt will automatically guess the file type by its extension and use the
-   * appropriate pre-processor. You will still need to install the required
-   * loader if you need to use them.
-   *
-   * @example
-   * ```js
-   * css: [
-   *   // Load a Node.js module directly (here it's a Sass file).
-   *   'bulma',
-   *   // CSS file in the project
-   *   '@/assets/css/main.css',
-   *   // SCSS file in the project
-   *   '@/assets/css/main.scss'
-   * ]
-   * ```
-   * @type {string[]}
-   * @version 2
-   * @version 3
-   */
-  css: {
-    $resolve: val => (val ?? []).map((c: any) => c.src || c)
-  },
 
   /**
    * An object where each key name maps to a path to a layout .vue file.
    *
    * Normally, there is no need to configure this directly.
    * @type {Record<string, string>}
-   * @version 2
    */
   layouts: {},
 
@@ -339,7 +139,6 @@ export default defineUntypedSchema({
    *
    * Normally, there is no need to configure this directly.
    * @type {string}
-   * @version 2
    */
   ErrorPage: null,
 
@@ -347,7 +146,6 @@ export default defineUntypedSchema({
    * Configure the Nuxt loading progress bar component that's shown between
    * routes. Set to `false` to disable. You can also customize it or create
    * your own component.
-   * @version 2
    */
   loading: {
     /** CSS color of the progress bar. */
@@ -384,7 +182,6 @@ export default defineUntypedSchema({
    * configuration. The name can refer to an indicator from [SpinKit](https://tobiasahlin.com/spinkit/)
    * or a path to an HTML template of the indicator source code (in this case, all the
    * other options will be passed to the template).
-   * @version 2
    */
   loadingIndicator: {
     $resolve: async (val, get) => {
@@ -408,7 +205,6 @@ export default defineUntypedSchema({
    *
    * @see [vue@2 documentation](https://v2.vuejs.org/v2/guide/transitions.html)
    * @see [vue@3 documentation](https://vuejs.org/guide/built-ins/transition-group.html#enter-leave-transitions)
-   * @version 2
    */
   pageTransition: {
     $resolve: async (val, get) => {
@@ -431,8 +227,6 @@ export default defineUntypedSchema({
    * to the `<Transition>` component that will wrap your layouts.
    *
    * @see [vue@2 documentation](https://v2.vuejs.org/v2/guide/transitions.html)
-   * @see [vue@3 documentation](https://vuejs.org/guide/built-ins/transition-group.html#enter-leave-transitions)
-   * @version 2
    */
   layoutTransition: {
     $resolve: val => {
@@ -446,7 +240,6 @@ export default defineUntypedSchema({
 
   /**
    * You can disable specific Nuxt features that you do not want.
-   * @version 2
    */
   features: {
     /** Set to false to disable Nuxt vuex integration */
