+++ packages/schema/src/nuxt2/config/cli.ts
@@ -1,16 +1,16 @@
 import { defineUntypedSchema } from 'untyped'
 
 export default defineUntypedSchema({
-  /**
+  cli: {
+    /**
    * Add a message to the CLI banner by adding a string to this array.
    * @type {string[]}
-   * @version 2
    */
-  badgeMessages: [],
+    badgeMessages: [],
 
-  /**
-   * Change the color of the 'Nuxt.js' title in the CLI banner.
-   * @version 2
-   */
-  bannerColor: 'green'
+    /**
+     * Change the color of the 'Nuxt.js' title in the CLI banner.
+     */
+    bannerColor: 'green'
+  }
 })
