+++ packages/schema/src/nuxt2/config/index.ts
@@ -1,0 +1,40 @@
+
+import app from './app'
+import build from './build'
+import cli from './cli'
+import common from './common'
+import generate from './generate'
+import messages from './messages'
+import render from './render'
+import router from './router'
+import server from './server'
+
+/*
+Schema differences from Nuxt2:
+- transition => pageTransition
+- export => generate
+- gzip => compressor
+- Apply preset
+- render.etag.hash should be a function
+- deprecated devModules
+- set consola level to 0 if build.quite is true
+- Ad-hoc: loading-screen, components and telemtry
+- build.indicator and build.loadingScreen
+- build.crossorigin => render.crossorigin
+- render.csp.unsafeInlineCompatiblity => render.csp.unsafeInlineCompatibility
+- guards: rootDir:buildDir rootDir:generate.dir srcDir:buildDir srcDir:generate.dir
+- _publicPath (original value of publicPath)
+- options.build.babel.presets (array) warn @nuxtjs/babel-preset-app => @nuxt/babel-preset-app
+*/
+
+export default {
+  ...app,
+  ...common,
+  ...build,
+  ...cli,
+  ...generate,
+  ...messages,
+  ...render,
+  ...router,
+  ...server
+}
