+++ packages/schema/src/nuxt2/hooks.ts
@@ -1,0 +1,126 @@
+import type { Server as HttpServer, IncomingMessage, ServerResponse } from 'node:http'
+import type { Server as HttpsServer } from 'node:https'
+import type { Compiler, Configuration, Stats } from 'webpack'
+import type { Nuxt2Config, Nuxt2Options } from './config'
+
+type HookResult = Promise<void> | void
+type ModuleContainer = any
+
+// TODO
+type Builder = any
+type Generator = any
+type Server = any
+type Nuxt = any
+type NuxtTemplate = any
+
+type WatchEvent = 'add' | 'addDir' | 'change' | 'unlink' | 'unlinkDir'
+
+interface PreloadFile {
+  asType: 'script' | 'style' | 'font'
+  extension: string
+  file: string
+  fileWithoutQuery: string
+}
+
+type RenderResult = {
+  html: string
+  cspScriptSrcHashes: string[]
+  error: any
+  redirected: boolean
+  preloadFiles: PreloadFile[]
+}
+
+export interface Nuxt2Hooks {
+  // @nuxt/builder
+  'build:before':
+  (builder: Builder, buildOptions: Nuxt2Options['build']) => HookResult
+  'builder:prepared': (builder: Builder, buildOptions: Nuxt2Options['build']) => HookResult
+  'builder:extendPlugins': (plugins: Nuxt2Options['plugins']) => HookResult
+  'build:templates': (templates: {
+    templateVars: Record<string, any>,
+    templatesFiles: NuxtTemplate[],
+    resolve: (...args: string[]) => string
+  }) => HookResult
+  'build:extendRoutes': (routes: any[], resolve: (...args: string[]) => string) => HookResult
+  'build:done': (builder: Builder) => HookResult
+  'watch:restart': (event: { event: string, path: string }) => HookResult
+  // 'watch:fileChanged': (builder: Builder, fileName: string) => HookResult
+  'builder:watch': (event: WatchEvent, path: string) => HookResult
+
+  // @nuxt/cli
+  'generate:cache:ignore': (ignore: string[]) => HookResult
+  'config': (options: Nuxt2Config) => HookResult
+  'run:before': (options: { argv: string[], cmd: { name: string, usage: string, description: string, options: Record<string, any> }, rootDir: string }) => HookResult
+
+  // @nuxt/core
+  'ready': (nuxt: Nuxt) => HookResult
+  'close': (nuxt: Nuxt) => HookResult
+  'modules:before': (moduleContainer: ModuleContainer, modules?: any[]) => HookResult
+  'modules:done': (moduleContainer: ModuleContainer) => HookResult
+  'webpack:done': () => HookResult
+
+  // @nuxt/server
+  'render:before': (server: Server, renderOptions: Nuxt2Options['render']) => HookResult
+  'render:setupMiddleware': (app: any) => HookResult
+  'render:errorMiddleware': (app: any) => HookResult
+  'render:done': (server: Server) => HookResult
+  'listen': (listenerServer: HttpServer | HttpsServer, listener: any) => HookResult
+  'server:nuxt:renderLoading': (req: IncomingMessage, res: ServerResponse) => HookResult
+  'render:route': (url: string, result: RenderResult, context: any) => HookResult
+  'render:routeDone': (url: string, result: RenderResult, context: any) => HookResult
+  'render:beforeResponse': (url: string, result: RenderResult, context: any) => HookResult
+
+  // @nuxt/vue-renderer
+  'render:resourcesLoaded': (resources: any) => HookResult
+  'vue-renderer:context': (renderContext: any) => HookResult
+  'vue-renderer:spa:prepareContext': (renderContext: any) => HookResult
+  'vue-renderer:spa:templateParams': (templateParams: Record<string, any>) => HookResult
+  'vue-renderer:ssr:prepareContext': (renderContext: any) => HookResult
+  'vue-renderer:ssr:context': (renderContext: any) => HookResult
+  '_render:context': (nuxt: Nuxt) => HookResult
+  'render:routeContext': (renderContext: any) => HookResult
+  'vue-renderer:ssr:csp': (cspScriptSrcHashes: string[]) => HookResult
+  'vue-renderer:ssr:templateParams': (templateParams: Record<string, any>, renderContext: any) => HookResult
+
+  // @nuxt/webpack
+  'webpack:config': (webpackConfigs: Configuration[]) => HookResult
+  'webpack:devMiddleware': (middleware: (req: IncomingMessage, res: ServerResponse, next: (err?: any) => any) => any) => HookResult
+  'webpack:hotMiddleware': (middleware: (req: IncomingMessage, res: ServerResponse, next: (err?: any) => any) => any) => HookResult
+  'server:devMiddleware': (middleware: (req: IncomingMessage, res: ServerResponse, next: (err?: any) => any) => any) => HookResult
+  'build:compile': (options: { name: string, compiler: Compiler }) => HookResult
+  'build:compiled': (options: { name: string, compiler: Compiler, stats: Stats }) => HookResult
+  'build:resources': (mfs?: Compiler['outputFileSystem']) => HookResult
+  'bundler:change': (shortPath: string) => void
+  'bundler:error': () => void
+  'bundler:done': () => void
+  'bundler:progress': (statesArray: any[]) => void
+
+  // @nuxt/generator
+  'generate:before': (generator: Generator, generateOptions: Nuxt2Options['generate']) => HookResult
+  'generate:distRemoved': (generator: Generator) => HookResult
+  'generate:distCopied': (generator: Generator) => HookResult
+  'generate:route': ({ route, setPayload }: { route: any, setPayload: any }) => HookResult
+  'generate:page': (page: {
+    route: any,
+    path: string,
+    html: string,
+    exclude: boolean,
+    errors: string[]
+  }) => HookResult
+  'generate:routeCreated': ({ route, path, errors }: { route: any, path: string, errors: any[] }) => HookResult
+  'generate:extendRoutes': (routes: any[]) => HookResult
+  'generate:routeFailed': ({ route, errors }: { route: any, errors: any[] }) => HookResult
+  'generate:manifest': (manifest: any, generator: Generator) => HookResult
+  'generate:done': (generator: Generator, errors: any[]) => HookResult
+
+  'export:before': (generator: Generator) => HookResult
+  'export:distRemoved': (generator: Generator) => HookResult
+  'export:distCopied': (generator: Generator) => HookResult
+  'export:route': ({ route, setPayload }: { route: any, setPayload: any }) => HookResult
+  'export:routeCreated': ({ route, path, errors }: { route: any, path: string, errors: any[] }) => HookResult
+  'export:extendRoutes': ({ routes }: { routes: any[] }) => HookResult
+  'export:routeFailed': ({ route, errors }: { route: any, errors: any[] }) => HookResult
+  'export:done': (generator: Generator, { errors }: { errors: any[] }) => HookResult
+}
+
+export type NuxtHookName = keyof Nuxt2Hooks
