+++ packages/schema/src/nuxt2/index.ts
@@ -1,0 +1,4 @@
+export * from './config'
+
+// Schema
+export { default as NuxtConfigSchema } from './config/index'
